chore: moved from site.json to site.js
