+++ .eleventy.js
@@ -27,8 +27,7 @@ const htmlMinTransform = require("./src/transforms/html-min-transform.js");
 const parseTransform = require("./src/transforms/parse-transform.js");
 
 // Import data files
-const site = require("./src/_data/site.json");
-
+const site = require("./src/_data/site");
 module.exports = function (config) {
   // Filters
   config.addFilter("dateFilter", dateFilter);
